Cleaned up tests and moved helper variables to _mock.js
