+++ src/tests/build-queries.js
@@ -1,61 +1,10 @@
-const builder = require("../builder.js");
 const test = require("ava");
+const builder = require("../builder.js");
+const mock = require("./_mock.js");
 const chalk = require("chalk");
 const log = console.log;
 
-const sampleMutation = `mutation {
-  createContent(
-    markup: $markup
-    raw: $raw
-  ) {
-    markup
-    raw
-  }
-}`;
-
-const queryVariablesArray = [
-  {
-    markup: "markup1",
-    raw: "raw1"
-  },
-  {
-    markup: "markup2",
-    raw: "raw2"
-  },
-  {
-    markup: "markup3",
-    raw: "raw3"
-  }
-];
-
-const expectedQueries = `mutation {
-  createContent(
-    markup: "markup1"
-    raw: "raw1"
-  ) {
-    markup
-    raw
-  }
-},mutation {
-  createContent(
-    markup: "markup2"
-    raw: "raw2"
-  ) {
-    markup
-    raw
-  }
-},mutation {
-  createContent(
-    markup: "markup3"
-    raw: "raw3"
-  ) {
-    markup
-    raw
-  }
-}`;
-
 test("queries build", t => {
-  let qs = builder(sampleMutation, queryVariablesArray);
-  let queries = String(qs);
-  t.is(queries, expectedQueries);
+  let queries = builder(mock.template, mock.variables);
+  t.is(String(queries), String(mock.batchQuery));
 });
